tests/wpt/tests/css/css-align/parsing/grid-gap-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing grid-gap with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align/#grid-gap-legacy">
<link rel="author" title="Takuya Kurimoto" href="mailto:takuya004869@gmail.com">
<meta name="assert" content="grid-gap supports the full grammar '<grid-row-gap> <grid-column-gap>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("grid-gap", "normal normal", "normal");
test_valid_value("grid-gap", "10px 10px", "10px");
test_valid_value("grid-gap", "20% 20%", "20%");
test_valid_value("grid-gap", "calc(20% + 10px) calc(20% + 10px)", "calc(20% + 10px)");

test_valid_value("grid-gap", "normal 10px");
test_valid_value("grid-gap", "10px 20%");
test_valid_value("grid-gap", "20% calc(20% + 10px)");
test_valid_value("grid-gap", "calc(20% + 10px) 0px");
test_valid_value("grid-gap", "0px normal");
</script>
</body>
</html>
